requirements.txt
streamlit
pillow
pytesseract
reportlab
numpy
easyocr
